requirements.txt
Django==4.2.3
djangorestframework==3.14.0
gunicorn==21.2.0
ipython==8.14.0
psycopg2==2.9.6
requests==2.31.0